Test updating image
